perf: New enable_on_starter attribute
If we want an event to count/sample only inside a starter/stopper
pair of event boundaries, we don't want it to start when it
is scheduled in context switch or when it is activated. We
want it to be set in the CPU but in a paused state, ready to be
resumed only when his starter fires an event.

The enable_on_starter attribute solves this problem by putting the
event in a paused state from the beginning until the starter activates
it.

Signed-off-by: Frederic Weisbecker <fweisbec@gmail.com>
Cc: Ingo Molnar <mingo@elte.hu>
Cc: Peter Zijlstra <a.p.zijlstra@chello.nl>
Cc: Arnaldo Carvalho de Melo <acme@redhat.com>
Cc: Paul Mackerras <paulus@samba.org>
Cc: Stephane Eranian <eranian@google.com
Cc: Steven Rostedt <rostedt@goodmis.org>
Cc: Masami Hiramatsu <masami.hiramatsu.pt@hitachi.com>
Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: Hitoshi Mitake <mitake@dcl.info.waseda.ac.jp>